Much better logic to return new call state using bit shifting and logical AND

+++ src/constants.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <cstdint>
 
 const int SCREEN_WIDTH = 1500;
 const int SCREEN_HEIGHT = 750;
@@ -12,4 +13,4 @@ const int GRID_COLS = (int)(SCREEN_WIDTH / CELL_SIZE);
 
 const int GRID_FEATHER = 1;
 
-const unsigned long RULE = 30;
+const uint8_t RULE = 30;
